Move FAB related functions to fab provider
